Doman: SRP mukainen muutos lukuvinkkipalvelu luokkaan, jotta luokalla vain yksi vastuu. URLPalvelu eritetty luokasta
